+++ opencascade-hs/src/OpenCascade/StlAPI/Types.hs
@@ -2,6 +2,8 @@
 {-# LANGUAGE EmptyDataDecls #-}
 module OpenCascade.StlAPI.Types 
 ( Writer
+, Reader
 ) where
 
 data Writer
+data Reader
